# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API token authentication. Use format: \"Bearer \u003cyour-api-token\u003e\"",
        "title": "Bible Notes API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "shuvoedward@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/v1",
    "paths": {
...
        "/v1/highlights": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new Bible verse highlight with color and position details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "highlights"
                ],
                "summary": "Create a new highlight",
                "parameters": [
                    {
                        "description": "Highlight data (user_id will be set automatically from auth)",
                        "name": "highlight",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/shuvoedward_Bible_project_internal_data.Highlight"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created highlight",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "highlight": {
                                    "$ref": "#/definitions/shuvoedward_Bible_project_internal_data.Highlight"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid JSON or request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "object",
                                "additionalProperties": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/highlights/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete an existing highlight owned by the authenticated user",
                "tags": [
                    "highlights"
                ],
                "summary": "Delete highlight",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Highlight ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully deleted highlight"
                    },
                    "400": {
                        "description": "Invalid highlight ID",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Highlight not found or unauthorized",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update the color of an existing highlight owned by the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "highlights"
                ],
                "summary": "Update highlight color",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Highlight ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Highlight color update",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated highlight",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "highlights": {
                                    "type": "object",
                                    "properties": {
                                        "color": {
                                            "type": "string"
                                        },
                                        "id": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request (bad ID or JSON)",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Highlight not found or unauthorized",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "shuvoedward_Bible_project_internal_data.Highlight": {
            "type": "object",
            "properties": {
                "book": {
                    "type": "string"
                },
                "chapter": {
                    "type": "integer"
                },
                "color": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "end_offset": {
                    "type": "integer"
                },
                "end_verse": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "start_offset": {
                    "type": "integer"
                },
                "start_verse": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "security": [
        {
            "ApiKeyAuth": []
        }
    ]
}